legal_assistant-app/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
chromadb
sentence-transformers
transformers
torch
gTTS
PyPDF2
python-docx
python-multipart
tqdm
